Add an `iter` function
For aarch64, we can do quite a bit better than just calling the `find`
function repeatedly: we build a bitset of 64 bits where we've already
found if they match the set of bits we're looking for. We can then
efficently iterate over those set bits.

It may be possible to do something similar in the x86 simd
implementation.